// ui/src/api/gen/PullOutcome.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { PullResult } from "./PullResult";
import type { RemoteSummary } from "./RemoteSummary";

export type PullOutcome = { op: number, branch: string, 
/**
 * `origin/main`.
 */
upstream: string, server: RemoteSummary, result: PullResult, };